src.py
The function accepts a stream of patient info as a list of patients and creates a patient. It analyzes each patient and registers a risk score. The function then adds each patient to a score. With that, the function separates the patients into the patient queue and patient stack. Finally, the function returns the queue of all the patients ready to be treated in the order of highest to lowest overall risk score.